[SPARK-14841][SQL] Move SQLBuilder into sql/core
