+++ sql/core/src/test/scala/org/apache/spark/sql/catalyst/ExpressionToSQLSuite.scala
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.spark.sql.hive
+package org.apache.spark.sql.catalyst
 
 import scala.util.control.NonFatal
