+++ sql/core/src/test/scala/org/apache/spark/sql/catalyst/SQLBuilderTest.scala
@@ -15,16 +15,17 @@
  * limitations under the License.
  */
 
-package org.apache.spark.sql.hive
+package org.apache.spark.sql.catalyst
 
 import scala.util.control.NonFatal
 
 import org.apache.spark.sql.{DataFrame, Dataset, QueryTest}
 import org.apache.spark.sql.catalyst.expressions.Expression
 import org.apache.spark.sql.catalyst.plans.logical.LogicalPlan
-import org.apache.spark.sql.hive.test.TestHiveSingleton
+import org.apache.spark.sql.test.SharedSQLContext
 
-abstract class SQLBuilderTest extends QueryTest with TestHiveSingleton {
+
+abstract class SQLBuilderTest extends QueryTest with SharedSQLContext {
   protected def checkSQL(e: Expression, expectedSQL: String): Unit = {
     val actualSQL = e.sql
     try {
@@ -33,19 +34,19 @@ abstract class SQLBuilderTest extends QueryTest with TestHiveSingleton {
       case cause: Throwable =>
         fail(
           s"""Wrong SQL generated for the following expression:
-             |
+              |
              |${e.prettyName}
-             |
+              |
              |$cause
            """.stripMargin)
     }
   }
 
   protected def checkSQL(plan: LogicalPlan, expectedSQL: String): Unit = {
-    val generatedSQL = try new SQLBuilder(plan, hiveContext).toSQL catch { case NonFatal(e) =>
+    val generatedSQL = try new SQLBuilder(plan).toSQL catch { case NonFatal(e) =>
       fail(
         s"""Cannot convert the following logical query plan to SQL:
-           |
+            |
            |${plan.treeString}
          """.stripMargin)
     }
@@ -56,9 +57,9 @@ abstract class SQLBuilderTest extends QueryTest with TestHiveSingleton {
       case cause: Throwable =>
         fail(
           s"""Wrong SQL generated for the following logical query plan:
-             |
+              |
              |${plan.treeString}
-             |
+              |
              |$cause
            """.stripMargin)
     }
